# Performance/CPU enhancements on region capture forms and image editor