--- a/ShareX.ScreenCaptureLib/Forms/RegionCaptureForm.cs
+++ b/ShareX.ScreenCaptureLib/Forms/RegionCaptureForm.cs
@@ -620,7 +620,6 @@ namespace ShareX.ScreenCaptureLib
             idleTimer.Start();
         }
 
-
         private void MonitorKey(int index)
         {
             if (index == 0)