### Region Capture Form & Image Editor
Instead of invalidating the form repeatedly on every OnPaint (a high intensity infinite loop, which chews up an entire CPU core) this change throttles invalidate calls to a maximum of once every 16ms (~60fps) when there is user input, and once every 250ms (~4fps) when there is no user input and repainting is not necessary.

The Pause and Resume on lost/got focus events still work as they did before.

### Light & Transparent Region Capture**
By checking if the selected region has changed, we can avoid reprocessing/redrawing the screen, saving CPU usage too.

---

I realize these changes slightly affect the display of the dotted border, but I think the CPU reduction is worth it for users who use slow devices or rely on battery power. Let me know if you think any amendments are needed.