@Jaex 2019-02-14
You literally reverted all my changes to how it was before. Long years ago I was also using timers and it was giving choppy experience. Main problem with using timer is, it is not high resolution timer so it's interval can trigger with different timings like even if you set it to 15ms it can trigger 30ms later. So my aim was making it smooth as possible. Also not having high resolution timer is the reason why I couldn't add max fps option to limit it to for example max 100 fps with Thread.Sleep etc. too.

Also there is no such thing as rendering it only when necessary because there can be a lot of animated stuffs in screen which requires rendering all the time, for example ant selection lines which always moving. Sure you can calculate regions of all animated objects and only re draw that regions in screen but it can get very complicated when there is a lot of animated objects. But there is one optimization I could do which is when region capture/image editor window lose focus then it stops rendering. So if someone do alt tab to work on something else then image editor won't try to render for no reason.

Even I used your system before and know how choppy it can be, still gave it a shot and pulled your PR for testing and I was right. Magnifier following cursor is so choppy it feels like 10 fps animation with that random jumping etc. So it is not acceptable like this and if it were I would have keep it like this before.

For people who prefer lower cpu and battery usage, can use Region capture (Light) which also using timers like yours: https://github.com/ShareX/ShareX/blob/master/ShareX.ScreenCaptureLib/Forms/RegionCaptureLightForm.cs#L73

Or they can use proper image editor software which will render only when there is changes. But this is just region capture form modified to act like image editor with a lot of workarounds. If I wanted to make normal image editor instead of on screen annotation, I would have made rendering system like that too.